Setup proper CCSM hardhat project separate from the SRI app
