+++ examples/bri-3/ccsm/contracts/CcsmBpiStateAnchor.sol
@@ -1,5 +1,5 @@
 //SPDX-License-Identifier: UNLICENSED
-pragma solidity 0.8.20;
+pragma solidity 0.8.24;
 
 import '@openzeppelin/contracts/access/AccessControl.sol';
