+++ examples/bri-3/package.json
@@ -41,7 +41,6 @@
     "@nestjs/schedule": "^3.0.1",
     "@nestjs/swagger": "^7.3.1",
     "@noble/curves": "^1.2.0",
-    "@openzeppelin/contracts": "^5.0.2",
     "@personaelabs/spartan-ecdsa": "^2.1.4",
     "@prisma/client": "^5.2.0",
     "circomlib": "^2.0.5",
@@ -58,7 +57,6 @@
     "ethr-did-resolver": "^10.1.0",
     "ew-precise-proofs-js": "^1.2.0",
     "fixed-merkle-tree": "^0.7.3",
-    "hardhat": "^2.17.2",
     "jose": "^4.13.1",
     "logplease": "^1.2.15",
     "merkletreejs": "^0.3.9",
@@ -77,7 +75,6 @@
     "@nestjs/cli": "^9.0.0",
     "@nestjs/schematics": "^9.0.0",
     "@nestjs/testing": "^9.0.0",
-    "@nomicfoundation/hardhat-ethers": "^3.0.6",
     "@types/bcrypt": "^5.0.0",
     "@types/circomlibjs": "^0.1.4",
     "@types/cron": "^2.0.1",
