+++ examples/bri-3/src/bri/ccsm/ccsm.module.ts
@@ -1,0 +1,15 @@
+import { Module } from '@nestjs/common';
+import { CcsmStorageAgent } from './agents/ccsmStorage.agent';
+import { EthereumService } from './services/ethereum.service';
+
+@Module({
+  providers: [
+    CcsmStorageAgent,
+    {
+      provide: 'IBlockchainService',
+      useClass: EthereumService,
+    },
+  ],
+  exports: [CcsmStorageAgent, 'IBlockchainService'],
+})
+export class CcsmModule {}
