+++ examples/bri-3/src/bri/ccsm/services/ccsm.interface.ts
