+++ examples/bri-3/src/bri/vsm/capabilites/executeVsmCycle/executeVsmCycleCommand.handler.ts
@@ -7,7 +7,7 @@ import { WorkflowStorageAgent } from '../../../workgroup/workflows/agents/workfl
 import { WorkstepStorageAgent } from '../../../workgroup/worksteps/agents/workstepsStorage.agent';
 import { ExecuteVsmCycleCommand } from './executeVsmCycle.command';
 import { WorkstepExecutedEvent } from '../handleWorkstepEvents/workstepExecuted.event';
-import { CcsmStorageAgent } from '../../../zeroKnowledgeProof/agents/ccsmStorage.agent';
+import { CcsmStorageAgent } from '../../../ccsm/agents/ccsmStorage.agent';
 
 @CommandHandler(ExecuteVsmCycleCommand)
 export class ExecuteVsmCycleCommandHandler
