--- examples/bri-3/src/bri/zeroKnowledgeProof/services/blockchain/blockchain.interface.ts
@@ -1,6 +1,0 @@
-import { Contract } from 'ethers';
-export interface IBlockchainService {
-  deployContract(contractName: string): Promise<void>;
-  connectToContract(contractName: string): Promise<Contract>;
-  storeAnchorHash(contractName: string, anchorHash: string): Promise<void>;
-}
