--- examples/bri-3/src/bri/zeroKnowledgeProof/services/blockchain/ethereum/contracts/Ccsm.sol
@@ -1,16 +1,0 @@
-//SPDX-License-Identifier: UNLICENSED
-pragma solidity 0.8.17;
-
-contract Ccsm {
-  mapping(string => bool) public anchorHashStore;
-
-  function setAnchorHash(string calldata _anchorHash) external {
-    anchorHashStore[_anchorHash] = true;
-  }
-
-  function getAnchorHash(
-    string calldata _anchorHash
-  ) external view returns (bool) {
-    return anchorHashStore[_anchorHash];
-  }
-}
