+++ examples/bri-3/src/bri/zeroKnowledgeProof/zeroKnowledgeProof.module.ts
@@ -1,8 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { LoggingModule } from '../../shared/logging/logging.module';
-import { CcsmStorageAgent } from './agents/ccsmStorage.agent';
-import { EthereumService } from './services/blockchain/ethereum/ethereum.service';
 import { CircuitInputsParserService } from './services/circuit/circuitInputsParser/circuitInputParser.service';
 import { SnarkjsCircuitService } from './services/circuit/snarkjs/snarkjs.service';
 
@@ -10,22 +8,12 @@ import { SnarkjsCircuitService } from './services/circuit/snarkjs/snarkjs.servic
   imports: [CqrsModule, LoggingModule],
 
   providers: [
-    CcsmStorageAgent,
     CircuitInputsParserService,
     {
       provide: 'ICircuitService',
       useClass: SnarkjsCircuitService,
     },
-    {
-      provide: 'IBlockchainService',
-      useClass: EthereumService,
-    },
   ],
-  exports: [
-    'ICircuitService',
-    'IBlockchainService',
-    CcsmStorageAgent,
-    CircuitInputsParserService,
-  ],
+  exports: ['ICircuitService', CircuitInputsParserService],
 })
 export class ZeroKnowledgeProofModule {}
