+++ examples/bri-3/tsconfig.json
@@ -23,5 +23,5 @@
   },
   "ts-node": {
     "compiler": "ttypescript"
-},
+  }
 }
